--fix:: Ensure Enter key triggers post submission in PostModal and -- Refactor:: Organize community page componenents
