+++ src/Component/Community/NewRule.jsx
